javadocs and a typo in project name
