--- instant-messanger/build.gradle
@@ -1,24 +1,0 @@
-//create a single Jar with all dependencies
-jar {
-    from {
-        configurations.compile.collect {
-            it.isDirectory() ? it : zipTree(it)
-        }
-    }
-    manifest {
-        attributes 'Main-Class': 'ru.mit.spbau.antonpp.messanger.Application'
-    }
-}
-
-dependencies {
-
-    // lombok annotations
-    compileOnly "org.projectlombok:lombok:1.16.10"
-
-    // matchers for tests
-    testCompile group: 'org.hamcrest', name: 'hamcrest-core', version: '1.3'
-    testCompile group: 'org.hamcrest', name: 'hamcrest-library', version: '1.3'
-
-    // mocking
-    testCompile "org.mockito:mockito-core:1.+"
-}
