--- instant-messanger/src/main/java/ru/mit/spbau/antonpp/messanger/network/receive/MessageReceiver.java
@@ -1,63 +1,0 @@
-package ru.mit.spbau.antonpp.messanger.network.receive;
-
-import lombok.extern.slf4j.Slf4j;
-import ru.mit.spbau.antonpp.messanger.network.data.SignedMessage;
-
-import java.io.Closeable;
-import java.io.DataInputStream;
-import java.io.IOException;
-import java.net.Socket;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-
-/**
- * Class that is responsible for receiving messages.
- *
- * @author Anton Mordberg
- * @since 17.02.17
- */
-@Slf4j
-public class MessageReceiver implements Closeable {
-
-    private final ExecutorService listeningService = Executors.newSingleThreadExecutor();
-    private final ReceiverCallback callback;
-    private final ServerConnectionProvider provider;
-
-    public MessageReceiver(ReceiverCallback callback, ServerConnectionProvider provider) throws IOException {
-        this.callback = callback;
-        this.provider = provider;
-
-        listeningService.execute(() -> {
-            while (true) {
-                try {
-                    onConnection(provider.accept());
-                } catch (IOException e) {
-                    log.warn("Failed to accept new connection", e);
-                    break;
-                }
-            }
-        });
-    }
-
-    @Override
-    public void close() throws IOException {
-        listeningService.shutdownNow();
-        provider.close();
-    }
-
-    private void onConnection(Socket clientSocket) {
-        try {
-            final DataInputStream dos = new DataInputStream(clientSocket.getInputStream());
-            callback.onReceive(SignedMessage.builder()
-                    .sender(dos.readUTF())
-                    .text(dos.readUTF())
-                    .build());
-        } catch (IOException e) {
-            log.warn("Connection failed", e);
-        }
-    }
-
-    public interface ReceiverCallback {
-        void onReceive(SignedMessage msg);
-    }
-}
