+++ instant-messenger/src/main/java/ru/mit/spbau/antonpp/messenger/network/data/SignedMessage.java
@@ -1,0 +1,17 @@
+package ru.mit.spbau.antonpp.messenger.network.data;
+
+import lombok.Builder;
+import lombok.Data;
+
+/**
+ * Simple Message that has only two fields.
+ *
+ * @author Anton Mordberg
+ * @since 17.02.17
+ */
+@Data
+@Builder
+public class SignedMessage {
+    private final String sender;
+    private final String text;
+}
