+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = 'au-software-design'
 
-include ':lab-bash', ':instant-messanger'
+include ':lab-bash', ':instant-messenger'
